Added destructible block and ground pound player state. Added gem pickup to increase life. Added door to next level.

+++ GameObject.cpp
@@ -21,7 +21,7 @@ void GameObject::UpdateAll(GameState& gState)
 {
 	std::sort(s_vUpdateList.begin(), s_vUpdateList.end(), GameObject::UpdateOrder);
 
-	for (int n = 0; n < s_vUpdateList.size(); n++)
+	for (size_t n = 0; n < s_vUpdateList.size(); n++)
 	{
 		s_vUpdateList[n]->Update(gState);
 
@@ -34,7 +34,7 @@ void GameObject::DrawAll(GameState& gState)
 {
 	std::sort(s_vDrawList.begin(), s_vDrawList.end(), GameObject::DrawOrder);
 
-	for (int n = 0; n < s_vDrawList.size(); n++)
+	for (size_t n = 0; n < s_vDrawList.size(); n++)
 		s_vDrawList[n]->Draw(gState);
 }
 
@@ -66,7 +66,7 @@ int GameObject::GetObjectList(GameObject::Type type, std::vector< GameObject* >&
 
 void GameObject::DestroyAll()
 {
-	for (int n = 0; n < s_vDrawList.size(); n++)
+	for (size_t n = 0; n < s_vDrawList.size(); n++)
 		delete s_vUpdateList[n--];
 }
